parser_hh/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
django-cleanup==4.0.0
django-debug-toolbar==2.2
django-filter==2.2.0
djangorestframework==3.11.0
Faker==0.9.1
flack==1.2.2
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mixer==6.1.3
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.8
Werkzeug==1.0.0
